1. Convert tv_sec to qint64 because we couldn't guarantee that time_t is wide enough; 2. Version bump.

+++ src/vkgeoappdelegate.mm
@@ -29,7 +29,7 @@ static qint64 elapsedNanos()
     struct timespec elapsed_time;
 
     if (clock_gettime(CLOCK_MONOTONIC_RAW, &elapsed_time) == 0) {
-        return elapsed_time.tv_sec * 1000000000 + elapsed_time.tv_nsec;
+        return (qint64)elapsed_time.tv_sec * 1000000000 + elapsed_time.tv_nsec;
     } else {
         return 0;
     }
